urbanherbapi/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.0
Pillow==10.1.0
django-filter==23.5
drf-nested-routers==0.93.4
